Applied classpath patches
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@1682 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ core/src/classpath/org/org/omg/CORBA/PUBLIC_MEMBER.java
@@ -51,5 +51,5 @@ public interface PUBLIC_MEMBER
   /**
    * The flag, specifying that the value member is public.
    */
-  int value = 1;
+  short value = 1;
 }
